Move these tests from routes.test.ts to WorksLink
